requirements.txt
python=3.8.5
pip=20.3
pytorch=1.12.1
torchvision=0.13.1
numpy=1.23.1
gradio==3.16.2
albumentations==1.3.0
imageio==2.9.0
imageio-ffmpeg==0.4.2
pytorch-lightning==1.5.0
omegaconf==2.1.1
einops==0.3.0
transformers==4.19.2
webdataset==0.2.5
open-clip-torch
torchmetrics==0.6.0
timm==0.6.12
addict==2.4.0
yapf==0.32.0
prettytable==3.6.0
safetensors
astroid
cycler
isort
kiwisolver
lazy-object-proxy
matplotlib
mccabe
protobuf
pycocotools
pylint
pyparsing
python-dateutil
scipy
six
tensorboardX
tqdm
typed-ast
wrapt
cython
cffi
easydict
Pillow>=6.2.2
yacs
pyyaml
visdom
bs4
html5lib
ninja
lightning
# opencv-python
# opencv-contrib-python
